# Conflicts:
#	EList.Services.Impl/MediaService.cs

+++ EList.Services.Interfaces/IEventsRatingService.cs
@@ -8,7 +8,7 @@ namespace EList.Services.Interfaces
     {
         Task<CommandResult<Guid>> VoteAsync(EventsRatingItem request);
         Task<CommandResult<EventRating>> GetEventRatingAsync(Guid eventId, EventRatingType eventRatingType, int? pageIndex, int? pageSize);
-        Task<CommandResult<int?>> GetOrganizatorRatingAsync(Guid accountId);
+        Task<CommandResult<double?>> GetOrganizatorRatingAsync(Guid accountId);
         Task<CommandResult> DeleteAsync(Guid itemId);
     }
 }
